+++ R/cv_ammif.R
@@ -89,7 +89,7 @@ cv_ammif <- function(.data, env, gen, rep, resp, nboot = 200, block, design = "R
                     Y = {{resp}})%>%
       mutate(across(1:3, as.factor))
     RMSPDres <- data.frame(RMSPD = matrix(0, nboot, 1))
-    data <- tibble::rowid_to_column(data)
+    data <- add_row_id(data)
     Nenv <- length(unique(data$ENV))
     Ngen <- length(unique(data$GEN))
     Nbloc <- length(unique(data$REP))
@@ -114,7 +114,7 @@ cv_ammif <- function(.data, env, gen, rep, resp, nboot = 200, block, design = "R
       df_test <-
         data.frame(
           test[which(test$test == TRUE),] %>%
-            remove_cols(rowid,REP, test) %>%
+            remove_cols(row_id,REP, test) %>%
             rename(n = Y)
         )
       message(paste0(capture.output(df_test), collapse = "\n"))
@@ -137,9 +137,9 @@ cv_ammif <- function(.data, env, gen, rep, resp, nboot = 200, block, design = "R
           modeling <- data %>%
             dplyr::group_by(ENV, GEN) %>%
             dplyr::sample_n(nrepval, replace = FALSE) %>%
-            arrange(rowid) %>%
+            arrange(row_id) %>%
             as.data.frame()
-          rownames(modeling) <- modeling$rowid      }
+          rownames(modeling) <- modeling$row_id      }
         if (design == "RCBD") {
           tmp <- split_factors(data, ENV, keep_factors = TRUE, verbose = FALSE)
           modeling <- do.call(rbind, lapply(tmp, function(x) {
@@ -148,9 +148,9 @@ cv_ammif <- function(.data, env, gen, rep, resp, nboot = 200, block, design = "R
               dplyr::group_by(GEN) %>%
               dplyr::filter(REP %in% X2)
           })) %>% as.data.frame()
-          rownames(modeling) <- modeling$rowid
+          rownames(modeling) <- modeling$row_id
         }
-        testing <- anti_join(data, modeling, by = c("ENV", "GEN", "REP", "Y", "rowid")) %>%
+        testing <- anti_join(data, modeling, by = c("ENV", "GEN", "REP", "Y", "row_id")) %>%
           arrange(ENV, GEN) %>%
           as.data.frame()
         MEDIAS <-
@@ -215,7 +215,7 @@ cv_ammif <- function(.data, env, gen, rep, resp, nboot = 200, block, design = "R
                     BLOCK = {{block}},
                     Y = {{resp}})
     RMSPDres <- data.frame(RMSPD = matrix(0, nboot, 1))
-    data <- tibble::rowid_to_column(data)
+    data <- add_row_id(data)
     Nenv <- length(unique(data$ENV))
     Ngen <- length(unique(data$GEN))
     Nbloc <- length(unique(data$REP))
@@ -240,7 +240,7 @@ cv_ammif <- function(.data, env, gen, rep, resp, nboot = 200, block, design = "R
       df_test <-
         data.frame(
           test[which(test$test == TRUE),] %>%
-            remove_cols(rowid,REP, test) %>%
+            remove_cols(row_id,REP, test) %>%
             rename(n = Y)
         )
       message(paste0(capture.output(df_test), collapse = "\n"))
@@ -263,18 +263,18 @@ cv_ammif <- function(.data, env, gen, rep, resp, nboot = 200, block, design = "R
           modeling <- data %>%
             dplyr::group_by(ENV, GEN) %>%
             dplyr::sample_n(nrepval, replace = FALSE) %>%
-            arrange(rowid) %>%
+            arrange(row_id) %>%
             as.data.frame()
-          rownames(modeling) <- modeling$rowid      }
+          rownames(modeling) <- modeling$row_id      }
         if (design == "RCBD") {
           tmp <- split_factors(data, ENV, keep_factors = TRUE, verbose = FALSE)
           modeling <- do.call(rbind, lapply(tmp, function(x) {
             X2 <- sample(unique(data$REP), nrepval, replace = FALSE)
             x %>% dplyr::group_by(GEN) %>% dplyr::filter(REP %in% X2)
           })) %>% as.data.frame()
-          rownames(modeling) <- modeling$rowid
+          rownames(modeling) <- modeling$row_id
         }
-        testing <- anti_join(data, modeling, by = c("ENV", "GEN", "REP", "BLOCK", "Y", "rowid")) %>%
+        testing <- anti_join(data, modeling, by = c("ENV", "GEN", "REP", "BLOCK", "Y", "row_id")) %>%
           arrange(ENV, GEN) %>%
           as.data.frame()
         MEDIAS <-