+++ R/utilities.R
@@ -8,7 +8,7 @@
 #'   \code{all_title_case()}, \code{stract_number()}, \code{stract_string()},
 #'   \code{remove_strings()}, and \code{tidy_strings()} \code{...} are the
 #'   variables to apply the function. If no variable is informed, the function
-#'   will be applied to all non-numeric variables in \code{.data}.
+#'   will be applied to all non-numeric variables in `.data`.
 #' @param digits The number of significant figures.
 #' @param pattern A string to be matched. Regular Expression Syntax is also
 #'   allowed.
@@ -352,13 +352,15 @@ tidy_strings <- function(.data, ..., sep = "_"){
 #' * \code{add_cols()}: Add one or more columns to an existing data frame. If
 #' specified \code{.before} or \code{.after} columns does not exist, columns are
 #' appended at the end of the data. Return a data frame with all the original
-#' columns in \code{.data} plus the columns declared in \code{...}. In
-#' \code{add_cols()} columns in \code{.data} are available for the expressions.
+#' columns in `.data` plus the columns declared in \code{...}. In
+#' \code{add_cols()} columns in `.data` are available for the expressions.
 #' So, it is possible to add a column based on existing data.
 #' * \code{add_rows()}: Add one or more rows to an existing data frame. If
 #' specified \code{.before} or \code{.after} rows does not exist, rows are
 #' appended at the end of the data. Return a data frame with all the original
-#' rows in \code{.data} plus the rows declared in \code{...} argument.
+#' rows in `.data` plus the rows declared in \code{...} argument.
+#' * \code{add_row_id()}: Add a column with the row id as the first column in
+#' `.data`.
 #' * \code{add_prefix()} and \code{add_suffix()} add prefixes and suffixes,
 #' respectively, in variable names selected in \code{...} argument.
 #' * \code{all_pairs()}: Get all the possible pairs between the levels of a
@@ -368,8 +370,13 @@ tidy_strings <- function(.data, ..., sep = "_"){
 #' * \code{colnames_to_title()}: Translate all column names to title case.
 #' * \code{column_exists()}: Checks if a column exists in a data frame. Return a
 #' logical value.
-#' * \code{columns_to_first()}: Move columns to first positions in \code{.data}.
-#' * \code{columns_to_last()}: Move columns to last positions in \code{.data}.
+#' * \code{columns_to_first()}: Move columns to first positions in `.data`.
+#' * \code{columns_to_last()}: Move columns to last positions in `.data`.
+#' * \code{columns_to_rownames()}: Move a column of `.data` to its row
+#' names.
+#' * \code{rownames_to_column()}: Move the row names of `.data` to a new
+#' column.
+#' * \code{remove_rownames()}: Remove the row names of `.data`.
 #' * \code{concatenate()}: Concatenate columns of a data frame. If \code{drop =
 #' TRUE} then the existing variables are dropped. If \code{pull = TRUE} then the
 #' concatenated variable is pull out to a vector. This is specially useful when
@@ -395,8 +402,8 @@ tidy_strings <- function(.data, ..., sep = "_"){
 #' @param .data A data frame
 #' @param ... The argument depends on the function used.
 #' * For \code{add_cols()} and \code{add_rows()} is name-value pairs. All values
-#' must have one element for each row in \code{.data} when using
-#' \code{add_cols()} or one element for each column in \code{.data} when using
+#' must have one element for each row in `.data` when using
+#' \code{add_cols()} or one element for each column in `.data` when using
 #' \code{add_rows()}. Values of length 1 will be recycled when using
 #' \code{add_cols()}.
 #'
@@ -408,7 +415,7 @@ tidy_strings <- function(.data, ..., sep = "_"){
 #'
 #' * For \code{columns_to_first()} and \code{columns_to_last()},  \code{...} is
 #' the column name or column index of the variable(s) to be moved to first or
-#' last in \code{.data}.
+#' last in `.data`.
 #'
 #' * For \code{remove_rows()} and \code{select_rows()}, \code{...} is an integer
 #' row value.
@@ -428,20 +435,21 @@ tidy_strings <- function(.data, ..., sep = "_"){
 #'   \code{add_suffix()}, respectively.
 #' @param new_var The name of the new variable containing the concatenated
 #'   values. Defaults to \code{new_var}.
+#' @param var Name of column to use for rownames.
 #' @param sep The separator to appear when using \code{concatenate()},
 #'   \code{add_prefix()}, or \code{add_suffix()}. Defaults to to \code{"_"}.
 #' @param drop Logical argument. If \code{TRUE} keeps the new variable
 #'   \code{new_var} and drops the existing ones. Defaults to \code{FALSE}.
 #' @param pull Logical argument. If \code{TRUE}, returns the last column (on the
 #'   assumption that's the column you've created most recently), as a vector.
-#' @param cols A quoted variable name to check if it exists in \code{.data}.
+#' @param cols A quoted variable name to check if it exists in `.data`.
 #' @param levels The levels of a factor or a numeric vector.
 #' @param offset Set it to \emph{n} to select the \emph{n}th variable from the
 #'   end (for \code{select_last_col()}) of from the begin (for
 #'   \code{select_first_col()})
 #' @md
 #' @author Tiago Olivoto \email{tiagoolivoto@@gmail.com}
-#' @importFrom  tibble add_column add_row
+#' @importFrom  tibble add_row
 #' @importFrom dplyr relocate across
 #' @export
 #' @examples
@@ -559,6 +567,15 @@ add_rows <- function(.data, ..., .before = NULL, .after = NULL){
 }
 #' @name utils_rows_cols
 #' @export
+add_row_id <- function(.data, var = "row_id"){
+  if(!has_class(.data, "data.frame")){
+    stop("The object '",  match.call()[[".data"]], "' must be a data frame.")
+  }
+  df <- .data
+  add_cols(df, {{var}} := seq_len(nrow(df)), .before = 1)
+}
+#' @name utils_rows_cols
+#' @export
 all_pairs <- function(.data, levels){
   levels <-
     get_levels(.data, {{levels}})
@@ -604,6 +621,33 @@ column_to_first <-function(.data, ...){
 #' @export
 column_to_last <-function(.data, ...){
   select_cols(.data, -c(!!!quos(...)), everything())
+}
+#' @name utils_rows_cols
+#' @export
+column_to_rownames <- function(.data, var = "rowname"){
+  df <-
+    as.data.frame(.data) %>%
+    remove_rownames()
+  rownames(df) <- df[[var]]
+  df[[var]] <- NULL
+  df
+}
+#' @name utils_rows_cols
+#' @export
+rownames_to_column <- function(.data, var = "rowname"){
+  df <- .data
+  if(var %in% colnames(df)){
+    stop("Variable '", var, "' already in data.", call. = FALSE)
+  }
+  df %>%
+    mutate(`:=`(!!var, rownames(df)), .before = 1) %>%
+    remove_rownames()
+}
+#' @name utils_rows_cols
+#' @export
+remove_rownames <- function(.data, ...){
+  rownames(.data) <- NULL
+  .data
 }
 #' @name utils_rows_cols
 #' @export
@@ -778,7 +822,7 @@ tidy_colnames <- function(.data, sep = "_"){
 #'
 #' * \strong{Useful functions for descriptive statistics. All of them work
 #' naturally with \code{\%>\%}, handle grouped data and multiple variables (all
-#' numeric variables from \code{.data} by default).}
+#' numeric variables from `.data` by default).}
 #'    - \code{av_dev()} computes the average absolute deviation.
 #'    - \code{ci_mean()} computes the confidence interval for the mean.
 #'    - \code{cv()} computes the coefficient of variation.
@@ -812,11 +856,11 @@ tidy_colnames <- function(.data, sep = "_"){
 #'  for grouping the data. Then the statistic required will be computed for all
 #'  numeric variables in the data. If no variables are informed in \code{...},
 #'  the statistic will be computed ignoring all non-numeric variables in
-#'  \code{.data}.
+#'  `.data`.
 #' * For the other statistics, \code{...} is a comma-separated of unquoted
 #'  variable names to compute the statistics. If no variables are informed in n
 #'  \code{...}, the statistic will be computed for all numeric variables in
-#'  \code{.data}.
+#'  `.data`.
 #' @param na.rm If \code{FALSE}, the default, missing values are removed with a
 #'   warning. If \code{TRUE}, missing values are silently removed.
 #' @param level The confidence level for the confidence interval of the mean.
@@ -1943,4 +1987,26 @@ coord_radar <- function (theta = "x", start = 0, direction = 1) {
   ggproto("CoordRadar", CoordPolar, theta = theta, r = r, start = start,
           direction = sign(direction),
           is_linear = function(coord) TRUE)
+}
+# check for data frames in lists
+has_df_in_list <- function(x){
+  !any(
+    is.na(
+      map(x, class) %>%
+        lapply(function(x){
+          match('data.frame', x)
+        })
+    )
+  )
+}
+# check if all are dataframes
+all_df_in_list <- function(x){
+  !all(
+    is.na(
+      map(x, class) %>%
+        lapply(function(x){
+          match('data.frame', x)
+        })
+    )
+  )
 }