+++ R/utils_sets.R
@@ -11,7 +11,8 @@
 #'   vector contains duplicates they will be discarded. If the list doesn't have
 #'   names the sets will be named as `"set_1"`, "`Set_2"`, `"Set_3"` and so on.
 #'   If vectors are given in `...`, the set names will be named with the names
-#'   of the objects provided.
+#'   of the objects provided. Data frames are also allowed, provided that common
+#'   column names exist.
 #' @param pairs Returns the pairwise unions of the sets? Defaults to `FALSE`.
 #' @return A vector showing the desired operation of the sets. If `pairs =
 #'   TRUE`, returns a list showing the pairwise operation of the sets.
@@ -29,6 +30,17 @@
 #' set_intersect(A, B, C)
 #' set_difference(B, C)
 #'
+#' # Operations with data frames
+#' # Add a row id for better understanding
+#' sets <- data_ge %>% add_row_id()
+#'
+#' set_1 <- sets[1:5,]
+#' set_2 <- sets[2:6,]
+#' set_3 <- sets[3:7,]
+#'
+#' set_intersect(set_1, set_2, set_3)
+#' set_difference(set_2, set_3)
+#' set_union(set_1, set_2, set_3)
 #' }
 #'
 
@@ -53,25 +65,26 @@ set_helper <- function(..., pairs = FALSE, fun = intersect){
         rlang::quo_get_expr(x)
       })
     )
-  if(is.list(c(...))){
-    sets <- as.list(...)
+  if(has_df_in_list(list(...))){
+    sets <- list(...)
+  } else{
+    if(is.list(c(...))){
+      sets <- as.list(...)
       if(is.null(names(sets))){
         names(sets) <- paste("set", seq_len(length(sets)), sep = "_")
       }
-  }else{
-    sets <- list(...)
-    if(is.null(names(sets))){
-      names(sets) <- set_name
+    }else{
+      sets <- list(...)
+      if(is.null(names(sets))){
+        names(sets) <- set_name
+      }
     }
   }
   if (length(sets) <= 1) {
     stop("The list must contain at least 2 vectors.", call. = FALSE)
   }
-  if (length(unique(sapply(sets, class))) != 1) {
+  if (!all_df_in_list(sets) & length(unique(sapply(sets, class))) != 1) {
     stop("Vectors must be in the same class.", call. = FALSE)
-  }
-  if (!(sapply(sets, class)[1] %in% c("integer", "numeric", "character"))) {
-    stop("The list must contain only integers, numerics or characters.", call. = FALSE)
   }
   if(is.null(names(sets))){
     names(sets) <- paste("set", seq_len(length(sets)), sep = "_")