+++ NEWS.md
@@ -2,6 +2,9 @@
 ## New functions
 * `get_levels_comb()` to get the combination of the levels of a factor.
 * `g_simula()` to simulate replicated genotype data.
+* `add_row_id()` to add a column as the row id.
+* `remove_rownames()` to remove the row names of a data frame.
+* `column_to_rownames()` and `rownames_to_column()` to deal with rownames.
 
 ## Minor improvements
 * New argument `sel.var()` in `corr_ci()` to filter correlations with a selected variable